Refactoring data.py into two separate scripts: db_dive.py and db_user.py.
