+++ octodeco/flaskr/user.py
@@ -1,10 +1,9 @@
 # Please see LICENSE.md
-import flask;
 from flask import (
-    Blueprint, request, render_template, redirect, flash, url_for
+    Blueprint, request, render_template, redirect, flash, url_for, abort
 )
 
-from . import data;
+from . import db_dive, db_user;
 
 bp = Blueprint('user', __name__, url_prefix='/user')
 
@@ -13,17 +12,17 @@ bp = Blueprint('user', __name__, url_prefix='/user')
 @bp.route('/info')
 def info():
     return render_template('user/info.html',
-                           divecount = data.get_dive_count(),
-                           diveinfos = data.get_all_dives(),
-                           user_details = data.get_user_details() );
+                           divecount = db_dive.get_dive_count(),
+                           diveinfos = db_dive.get_all_dives(),
+                           user_details = db_user.get_user_details());
 
 
 @bp.route('/update', methods = [ 'POST' ])
 def update():
     action = request.form.get('action');
     if action.startswith('Reset'):
-        data.user_reset_profile();
+        db_user.user_reset_profile();
         flash('User profile reset');
         return redirect(url_for('user.info'));
     else:
-        flask.abort(405);
+        abort(405);
